+++ bootstrap_cfn/fab_tasks.py
@@ -13,7 +13,10 @@ from bootstrap_cfn.config import ProjectConfig, ConfigParser
 from bootstrap_cfn.cloudformation import Cloudformation
 from bootstrap_cfn.iam import IAM
 from bootstrap_cfn.elb import ELB
+from bootstrap_cfn.r53 import R53
 from bootstrap_cfn.utils import tail
+from boto.route53.exception import DNSServerError
+import uuid
 
 
 # Default fab config. Set via the tasks below or --set
@@ -83,6 +86,24 @@ def application(application_name):
 
 
 @task
+def tag(tag):
+    """
+    Set a tag for the stack
+
+    Sets the environment variable 'tag'
+    This gets used to store a DNS entry to identify
+    multiple stacks with the same name.
+    e.g. you can tag a stack as active, or inactive,
+    green or blue etc.
+
+    Args:
+        tag(string): The string to set the
+        variable to
+    """
+    env.tag = str(tag).lower()
+
+
+@task
 def config(config_file):
     """
     Set the location of the project's YAML file
@@ -148,16 +169,81 @@ def user(username):
     env.user = username
 
 
-def get_stack_name():
+@task
+def swap_tags(tag1, tag2):
+    """
+    Swap two tagged stacks.
+
+    i.e. update the DNS text record which defines the
+    random suffix associated with a stack tag.
+    """
+    cfn_config = get_config()
+    r53_conn = get_connection(R53)
+    zone_name = cfn_config.data['master_zone']
+    zone_id = r53_conn.get_hosted_zone_id(zone_name)
+    legacy_name = "{0}-{1}".format(env.application, env.environment)
+    record1 = "stack.{0}.{1}".format(tag1, legacy_name)
+    record2 = "stack.{0}.{1}".format(tag2, legacy_name)
+    stack_suffix1 = r53_conn.get_record(zone_name, zone_id, record1, 'TXT')
+    stack_suffix2 = r53_conn.get_record(zone_name, zone_id, record2, 'TXT')
+    fqdn1 = "{0}.{1}".format(record1, zone_name)
+    fqdn2 = "{0}.{1}".format(record2, zone_name)
+    r53_conn.update_dns_record(zone_id, fqdn1, 'TXT', '"{0}"'.format(stack_suffix2))
+    r53_conn.update_dns_record(zone_id, fqdn2, 'TXT', '"{0}"'.format(stack_suffix1))
+
+
+def get_stack_name(new=False):
     """
     Get the name of the stack
 
     The name of the stack is a combination
     of the application and environment names
+    and a randomly generated suffix.
+
+    The env.tag dictates which randomly generated suffix
+    the default env.tag is 'active'
+
+    If new=True we generate a new stack_name and create the
+    dns records to retreive it in the future.
+
     """
-    if hasattr(env, 'stack_name'):
-        return env.stack_name
-    return "%s-%s" % (env.application, env.environment)
+    if hasattr(env, 'tag'):
+        tag = env.tag
+    else:
+        tag = 'active'
+        env.tag = tag
+    if not hasattr(env, 'stack_name') or new:
+        legacy_name = "{0}-{1}".format(env.application, env.environment)
+        # get_config needs a stack_name so this is a hack because we don't
+        # know it yet...
+        env.stack_name = 'temp'
+        cfn_config = get_config()
+        try:
+            r53_conn = get_connection(R53)
+            zone_name = cfn_config.data['master_zone']
+            zone_id = r53_conn.get_hosted_zone_id(zone_name)
+            record_name = "stack.{0}.{1}".format(tag, legacy_name)
+            if new:
+                stack_suffix = uuid.uuid4().__str__()[-8:]
+                record = "{0}.{1}".format(record_name, zone_name)
+                r53_conn.update_dns_record(zone_id, record, 'TXT', '"{0}"'.format(stack_suffix))
+            else:
+                stack_suffix = r53_conn.get_record(zone_name, zone_id, record_name, 'TXT')
+            if stack_suffix:
+                env.stack_name = "{0}-{1}".format(legacy_name, stack_suffix)
+            else:
+                env.stack_name = legacy_name
+        except KeyError:
+            logging.warn("No master_zone in yaml, unable to create/find DNS records for "
+                         "stack name, will fallback to legacy stack names: "
+                         "application-environment")
+            env.stack_name = legacy_name
+        except DNSServerError:
+            logging.warn("Couldn't find/create DNS entry for stack suffix, "
+                         "stack name, will fallback to legacy stack names: "
+                         "application-environment")
+            env.stack_name = legacy_name
+    return env.stack_name
 
 
 def _validate_fabric_env():
@@ -242,7 +328,7 @@ def cfn_create():
     specification will be generated and used to create a
     stack on AWS.
     """
-    stack_name = get_stack_name()
+    stack_name = get_stack_name(new=True)
     cfn_config = get_config()
 
     cfn = get_connection(Cloudformation)